requirements.txt
Flask
Flask-Injector
connexion[swagger-ui]
waitress